requirements/requirements-model.txt
matplotlib==3.8.0
seaborn==0.13.2
numpy==1.24.3
pandas==2.2.3
PyYAML==6.0.2
scikit_learn==1.3.0
